Make migrations. Setting admin of repository.
