+++ commits/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class CommitsConfig(AppConfig):
+    name = 'commits'
