+++ lit/urls.py
@@ -21,7 +21,7 @@ urlpatterns = [
     url(r'^docs/', include_docs_urls(title='Todo API', description='RESTful API for Todo')),
 
     url(r'^admin/', admin.site.urls),
-    url(r'^$', include('rest_framework.urls', namespace='rest_framework')),
+    url(r'^api-auth/', include('rest_framework.urls', namespace='rest_framework')),
     url(r'^', include('users.urls', namespace='users')),
     url(r'^', include('repositories.urls', namespace='repositories')),
 ]
