+++ repositories/admin.py
@@ -1,3 +1,11 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import Repository
+
+
+class RepositoryAdmin(admin.ModelAdmin):
+    list_filter = ['name']
+    search_fields = ['name']
+
+
+admin.site.register(Repository, RepositoryAdmin)
