+++ repositories/views.py
@@ -1,6 +1,7 @@
+from rest_framework import generics
+
 from repositories.models import Repository
 from repositories.serializers import RepositorySerializer
-from rest_framework import generics
 
 
 class RepositoryList(generics.ListCreateAPIView):
